Core: split inline layout into its parts
The 4300-line inline.rs becomes a module per concern: items, floats,
metrics, truncation, text_fit, outline, background, runs, decorations and
breaking, with the request/build entry points in mod.rs. Every public path
stays through re-exports.
